Introduce autoreset after frame processing
After a frame is processed a new frame is assumed to enter the VAD.
It is then necessary to reset the accumulator at the end of the frame so
that a new energy averaging is evaluated. This is accomplished by using
the counter_tick.

As soon as the counter_tick raises, the accumulator is reset, so the
comparator will output a logical '0' before the rising edge of the next
clock, when the vad_out register will be enabled (note that couter_tick
enables the clock signal vad_out input). As a result, the vad_out output
will be '0' regardless of the data input sequence.

In order to solve this problem, the voice_detected signal is delayed by
1 clock period by means of the flip flop voice_detected_delay_register.
This flip flop will keep the comparator output even if the accumulator
is reset. At the next clock rising edge, when the vad_out clock is
enabled, the vad_out input will be the delayed output of the comparator,
so the VAD output is correct.