Reduce bit number of abs representation
Since we approximate the representation of the absolute value of
negative numbers by removing the increment of the complement (see note
below), the maximum absolute value representation will be in
the range 0 to 2^15 - 1, thus we do not require the 16th bit to
represent 2^15.

Note:
Recall that if x is represented by X, -x is represented by compl(X) + 1.
If X is the 2 complement representation and x < 0, then X_MSB is '1'.
Then compl(X_MSB) is '0', so we can discard the MSB from the
representation for not being significant any more, provided that we
approximate the representation of -x with compl(X) (without the
increment).