Make the testbench interactive
Instead of creating a static testbench, it is possible to feed data into
the cases for the testbench through the stdin. This testbench design
allows for flexible testing without recompiling the vhdl tb design.
For running a different test it is sufficient to feed different data
through stdin at runtime.